==> RecSys2021_Challenge/01_training/stage1/giba/2-create-features.ipynb <==
# Copyright 2021 NVIDIA CORPORATION
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

import os
import time
os.environ["CUDA_VISIBLE_DEVICES"]="0,1,2,3"
import gc
import glob
import joblib
import cupy as cp
import pandas as pd
from tqdm.notebook import tqdm
from joblib import Parallel, delayed
import os, time
from tqdm.notebook import tqdm

targets = ['reply','retweet','retweet_comment','like']

files = glob.glob('/raid/kaggle/2021/recsys/input/train_proc3/*')
print( len(files), files[1])

start = time.time()
very_start = time.time()

import numpy as np
from datetime import datetime
import matplotlib.pyplot as plt
import cudf, cupy, time, rmm
cudf.__version__

from dask.distributed import Client, wait
from dask_cuda import LocalCUDACluster
from dask.utils import parse_bytes
import cudf
import dask_cudf

import os
import time
import glob
import gc

import cupy as cp          # CuPy is an implementation of NumPy-compatible multi-dimensional array on GPU
import cudf                # cuDF is an implementation of Pandas-like Dataframe on GPU
import rmm                 # library for pre-allocating memory on GPU
import dask                # dask is an open-source library to nateively scale Python on multiple workers/nodes
import dask_cudf           # dask_cudf uses dask to scale cuDF dataframes on multiple workers/nodes

import numpy as np
import xgboost as xgb

# More dask / dask_cluster related libraries to scale NVTabular
from dask_cuda import LocalCUDACluster
from dask.distributed import Client
from dask.distributed import wait
from dask.utils import parse_bytes
from dask.delayed import delayed

# NVTabular is the core library, we will use here for feature engineering/preprocessing on GPU
#import nvtabular as nvt
#from nvtabular.utils import device_mem_size
!rm -r dask-worker-space

cluster = LocalCUDACluster(
    local_directory='/raid/kaggle/',
    CUDA_VISIBLE_DEVICES = '0,1,2,3',
    device_memory_limit = 0.50,
)
client = Client(cluster)

client

path = '/raid/kaggle/2021/recsys/input/train_proc3'
df = dask_cudf.read_parquet(f'{path}/*.parquet', npartitions=253)
df.columns

%%time
targets = ['reply','retweet','retweet_comment','like']

features = [
    'hashtags', 'tweet_id', 'media', 'links', 'domains', 'tweet_type',
    'language', 'timestamp', 'a_user_id', 'a_follower_count',
    'a_following_count', 'a_is_verified', 'a_account_creation', 'b_user_id',
    'b_follower_count', 'b_following_count', 'b_is_verified',
    'b_account_creation', 'b_follows_a', 'reply', 'retweet',
    'retweet_comment', 'like', 'tw_len_token', 'tw_len_media',
    'tw_len_photo', 'tw_len_video', 'tw_len_gif', 'tw_len_quest',
    'tw_count_capital_words', 'tw_count_excl_quest_marks',
    'tw_count_special1', 'tw_count_hash', 'tw_last_quest', 'tw_len_retweet',
    'tw_len_rt', 'tw_count_at', 'tw_count_words', 'tw_count_char',
    'tw_rt_count_words', 'tw_rt_count_char', 'tw_original_user0',
    'tw_original_user1', 'tw_original_user2', 'tw_rt_user0', 'tw_word0',
    'tw_word1', 'tw_tweet', 'group', 'dt_day', 'dt_dow', 'dt_minute',
    'len_hashtags', 'len_links', 'len_domains',
    'decline'
]

path = '/raid/kaggle/2021/recsys/input/train_proc3'
df = dask_cudf.read_parquet(f'{path}/*.parquet', npartitions=253, columns=features )
_ = wait(df)
df.head()

a = df.shape
a[0].compute(),a[1]

df.columns

df['dt_day'].value_counts().compute()

dt = df.loc[df.dt_day < 18]
dt.to_parquet( 'week12' )

dt = df.loc[df.dt_day >= 18]
dt.to_parquet( 'week3' )
gc.collect()

!mkdir week3-train
!mkdir week3-valid
files = glob.glob('week3/*.parquet')
for i, fn in tqdm(enumerate(files)):
    df = pd.read_parquet(fn)
    df = df.sample(frac=1.0)
    sz = int(df.shape[0]*0.20)
    df.iloc[:sz].to_parquet( 'week3-valid/part.'+str(i)+'.parquet' )
    df.iloc[sz:].to_parquet( 'week3-train/part.'+str(i)+'.parquet' )

trainfiles = glob.glob('week12/*.parquet') + glob.glob('week3-train/*.parquet')
len(trainfiles)

df = dask_cudf.read_parquet(trainfiles, npartitions=253, columns=features )
_ = wait(df)
df['decline'] = df['decline']//7.19
df.head()

!rm -r TEMAPS_LOCAL
!mkdir TEMAPS_LOCAL

for target in targets:
    feats = ['b_user_id', 'tweet_type']
    dt = df.groupby(feats)[target].agg(['sum','count'])
    fname = 'te_' + target + '-' + '-'.join(feats)
    dt.columns = ['sums_'+fname,'counts_'+fname]
    for f in dt.columns:
        dt[f] = dt[f].astype(np.int32)

    dt.to_parquet('TEMAPS_LOCAL/'+fname+'.parquet')

%%time
dt0 = pd.read_parquet( 'TEMAPS_LOCAL/te_reply-b_user_id-tweet_type.parquet' )
dt1 = pd.read_parquet( 'TEMAPS_LOCAL/te_retweet-b_user_id-tweet_type.parquet' )
dt2 = pd.read_parquet( 'TEMAPS_LOCAL/te_retweet_comment-b_user_id-tweet_type.parquet' )
dt3 = pd.read_parquet( 'TEMAPS_LOCAL/te_like-b_user_id-tweet_type.parquet' )
dt0 = dt0.sort_index()
dt1 = dt1.sort_index()
dt2 = dt2.sort_index()
dt3 = dt3.sort_index()
dt0.shape, dt1.shape, dt2.shape, dt3.shape

dt0['counts_te_reply-b_user_id-tweet_type'] = dt0['counts_te_reply-b_user_id-tweet_type'].astype(np.int16)
dt0['sums_te_reply-b_user_id-tweet_type'] = dt0['sums_te_reply-b_user_id-tweet_type'].astype(np.int16)
dt0['sums_te_retweet-b_user_id-tweet_type'] = dt1['sums_te_retweet-b_user_id-tweet_type'].astype(np.int16)
dt0['sums_te_retweet_comment-b_user_id-tweet_type'] = dt2['sums_te_retweet_comment-b_user_id-tweet_type'].astype(np.int16)
dt0['sums_te_like-b_user_id-tweet_type'] = dt3['sums_te_like-b_user_id-tweet_type'].astype(np.int16)

dt0.to_parquet('TEMAPS_LOCAL/b_user_id-tweet_type.parquet')

dt = pd.read_parquet('TEMAPS_LOCAL/b_user_id-tweet_type.parquet')
print(dt.shape)
dt.head()

feats = ['b_user_id']
dt = df.groupby(feats)[targets].agg({
    "reply": ['sum'],
    "retweet": ['sum'],
    "retweet_comment": ['sum'],
    "like": ['sum','count'],
})
dt.columns = ['sums_b_user_reply','sums_b_user_retweet','sums_b_user_retweet_comment','sums_b_user_like','counts_b_user_id']
for f in dt.columns:
    dt[f] = dt[f].astype(np.int32)

dt.to_parquet('TEMAPS_LOCAL/te-b_user_id.parquet')

dt = pd.read_parquet('TEMAPS_LOCAL/te-b_user_id.parquet')
dt = dt.sort_index()
dt.to_parquet('TEMAPS_LOCAL/b_user_id.parquet')
dt

for target in targets:
    feats = ['a_user_id', 'tweet_type']
    dt = df.groupby(feats)[target].agg(['sum','count'])
    fname = 'te_' + target + '-' + '-'.join(feats)
    dt.columns = ['sums_'+fname,'counts_'+fname]
    for f in dt.columns:
        dt[f] = dt[f].astype(np.int32)

    dt.to_parquet('TEMAPS_LOCAL/'+fname+'.parquet')

dt0 = pd.read_parquet( 'TEMAPS_LOCAL/te_reply-a_user_id-tweet_type.parquet' )
dt1 = pd.read_parquet( 'TEMAPS_LOCAL/te_retweet-a_user_id-tweet_type.parquet' )
dt2 = pd.read_parquet( 'TEMAPS_LOCAL/te_retweet_comment-a_user_id-tweet_type.parquet' )
dt3 = pd.read_parquet( 'TEMAPS_LOCAL/te_like-a_user_id-tweet_type.parquet' )
dt0 = dt0.sort_index()
dt1 = dt1.sort_index()
dt2 = dt2.sort_index()
dt3 = dt3.sort_index()
dt0.shape, dt1.shape, dt2.shape, dt3.shape

dt0['counts_te_reply-a_user_id-tweet_type'] = dt0['counts_te_reply-a_user_id-tweet_type'].astype(np.int16)
dt0['sums_te_reply-a_user_id-tweet_type'] = dt0['sums_te_reply-a_user_id-tweet_type'].astype(np.int16)
dt0['sums_te_retweet-a_user_id-tweet_type'] = dt1['sums_te_retweet-a_user_id-tweet_type'].astype(np.int16)
dt0['sums_te_retweet_comment-a_user_id-tweet_type'] = dt2['sums_te_retweet_comment-a_user_id-tweet_type'].astype(np.int16)
dt0['sums_te_like-a_user_id-tweet_type'] = dt3['sums_te_like-a_user_id-tweet_type'].astype(np.int16)

dt0.to_parquet('TEMAPS_LOCAL/a_user_id-tweet_type.parquet')

feats = ['media','language','tweet_type','a_is_verified','b_is_verified','b_follows_a','tw_last_quest','decline','group']
dt = df.groupby(feats)[targets].agg({
    "reply": ['sum'],
    "retweet": ['sum'],
    "retweet_comment": ['sum'],
    "like": ['sum','count'],
})
fname = 'te-' + '-'.join(feats)
dt.columns = ['multi_reply','multi_retweet','multi_retweet_comment','multi_like','multi_counts']
for f in dt.columns:
    dt[f] = dt[f].astype(np.int32)
    
dt.to_parquet('TEMAPS_LOCAL/'+fname+'.parquet')

dt = pd.read_parquet('TEMAPS_LOCAL/te-media-language-tweet_type-a_is_verified-b_is_verified-b_follows_a-tw_last_quest-decline-group.parquet')
dt = dt.sort_index()
dt.to_parquet('TEMAPS_LOCAL/media-language-tweet_type-a_is_verified-b_is_verified-b_follows_a-tw_last_quest-decline-group.parquet')
dt

feats = ['tw_word0', 'tweet_type']
dt = df.groupby(feats)[targets].agg({
    "reply": ['sum'],
    "retweet": ['sum'],
    "retweet_comment": ['sum'],
    "like": ['sum','count'],
})
fname = 'te-' + '-'.join(feats)
dt.columns = ['word_reply','word_retweet','word_retweet_comment','word_like','word_counts']
for f in dt.columns:
    dt[f] = dt[f].astype(np.int32)

dt.to_parquet('TEMAPS_LOCAL/'+fname+'.parquet')

dt = pd.read_parquet('TEMAPS_LOCAL/te-tw_word0-tweet_type.parquet')
dt = dt.sort_index()
dt.to_parquet('TEMAPS_LOCAL/tw_word0-tweet_type.parquet')
print(dt.shape)
dt

feats = ['tw_original_user0', 'tweet_type']
dt = df.groupby(feats)[targets].agg({
    "reply": ['sum'],
    "retweet": ['sum'],
    "retweet_comment": ['sum'],
    "like": ['sum','count'],
})
fname = 'te-' + '-'.join(feats)
dt.columns = ['ouser0_reply','ouser0_retweet','ouser0_retweet_comment','ouser0_like','ouser0_counts']
for f in dt.columns:
    dt[f] = dt[f].astype(np.int32)

dt.to_parquet('TEMAPS_LOCAL/'+fname+'.parquet')

dt = pd.read_parquet('TEMAPS_LOCAL/te-tw_original_user0-tweet_type.parquet')
dt = dt.sort_index()
dt.to_parquet('TEMAPS_LOCAL/tw_original_user0-tweet_type.parquet')
print(dt.shape)
dt

feats = ['tw_original_user1', 'tweet_type']
dt = df.groupby(feats)[targets].agg({
    "reply": ['sum'],
    "retweet": ['sum'],
    "retweet_comment": ['sum'],
    "like": ['sum','count'],
})
fname = 'te-' + '-'.join(feats)
dt.columns = ['ouser1_reply','ouser1_retweet','ouser1_retweet_comment','ouser1_like','ouser1_counts']
for f in dt.columns:
    dt[f] = dt[f].astype(np.int32)

dt.to_parquet('TEMAPS_LOCAL/'+fname+'.parquet')

dt = pd.read_parquet('TEMAPS_LOCAL/te-tw_original_user1-tweet_type.parquet')
dt = dt.sort_index()
dt.to_parquet('TEMAPS_LOCAL/tw_original_user1-tweet_type.parquet')
print(dt.shape)
dt

feats = ['tw_rt_user0' ]
dt = df.groupby(feats)[targets].agg({
    "reply": ['sum'],
    "retweet": ['sum'],
    "retweet_comment": ['sum'],
    "like": ['sum','count'],
})
fname = 'te-' + '-'.join(feats)
dt.columns = ['rtuser0_reply','rtuser0_retweet','rtuser0_retweet_comment','rtuser0_like','rtuser0_counts']
for f in dt.columns:
    dt[f] = dt[f].astype(np.int32)

dt.to_parquet('TEMAPS_LOCAL/'+fname+'.parquet')

dt = pd.read_parquet('TEMAPS_LOCAL/te-tw_rt_user0.parquet')
dt = dt.sort_index()
dt.to_parquet('TEMAPS_LOCAL/tw_rt_user0.parquet')
print(dt.shape)
dt

df['a_follower_count'] = df['a_follower_count'] // 819
df['a_following_count'] = df['a_following_count'] // 712
df['b_follower_count'] = df['b_follower_count'] // 819
df['b_following_count'] = df['b_following_count'] // 712

(df['a_following_count']>100).mean().compute()

feats = ['a_follower_count', 'a_following_count','b_follower_count', 'b_following_count', 'tweet_type', 'language', 'b_follows_a']
dt = df.groupby(feats)[targets].agg({
    "reply": ['sum'],
    "retweet": ['sum'],
    "retweet_comment": ['sum'],
    "like": ['sum','count'],
})
fname = 'te-' + '-'.join(feats)
dt.columns = ['follow_reply','follow_retweet','follow_retweet_comment','follow_like','follow_counts']
for f in dt.columns:
    dt[f] = dt[f].astype(np.int32)

dt.to_parquet('TEMAPS_LOCAL/'+fname+'.parquet')

dt = pd.read_parquet('TEMAPS_LOCAL/te-a_follower_count-a_following_count-b_follower_count-b_following_count-tweet_type-language-b_follows_a.parquet')
dt = dt.sort_index()
dt.to_parquet('TEMAPS_LOCAL/a_follower_count-a_following_count-b_follower_count-b_following_count-tweet_type-language-b_follows_a.parquet')
print(dt.shape)
dt

features = [
    'media', 'tweet_type', 'language', 
    'a_user_id32', 'a_is_verified', 
    'b_user_id32', 'b_is_verified', 'b_follows_a',
    'reply', 'retweet',  'retweet_comment', 'like',
    'tw_original_user0', 'tw_original_user1', 'tw_original_user2',
    'tw_rt_user0',
    'tw_word0', 'tw_word1', 
    'tw_tweet',
    'group', 
    'dt_dow',
    'decline'
]

path = '/raid/kaggle/2021/recsys/input/train_proc3'
df = dask_cudf.read_parquet(f'{path}/*.parquet', npartitions=500, columns=features )
_ = wait(df)
df.head()

df['positive'] = 1*(df['reply']>0)+2*(df['retweet']>0)+4*(df['retweet_comment']>0)+8*(df['like']>0)
df['positive'].value_counts().compute()

feats = ['a_user_id32', 'b_user_id32']
for target in ['reply', 'retweet',  'retweet_comment', 'like']:
    dt = df[feats+[target,'positive']].loc[df.positive>0].groupby(feats)[target].agg('sum')
    fname = 'te-' + '-'.join(feats) + '_'+ target
    dt.columns = ['ab_'+target]
    for f in dt.columns:
        dt[f] = dt[f].astype(np.int32)
    dt.to_parquet('TEMAPS_LOCAL/'+fname+'.parquet')

feats = ['a_user_id32', 'b_user_id32']
dt = df[feats+['positive']].loc[df.positive>0].groupby(feats)['positive'].agg('count')
fname = 'te-' + '-'.join(feats) + '_count'
dt.columns = ['ab_count']
for f in dt.columns:
    dt[f] = dt[f].astype(np.int32)
dt.to_parquet('TEMAPS_LOCAL/'+fname+'.parquet')

print( dt.head() )
