==> RecSys2021_Challenge/01_training/stage1/giba/3-XGB-training.ipynb <==
# Copyright 2021 NVIDIA CORPORATION
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

import os
import time
#os.environ["CUDA_VISIBLE_DEVICES"]="2,3"
os.environ["CUDA_VISIBLE_DEVICES"]="0"
import gc
import glob
import cupy as cp
import pandas as pd
from tqdm.notebook import tqdm
from joblib import Parallel, delayed
import os, time
from tqdm.notebook import tqdm
import pickle

import numpy as np
from datetime import datetime
import matplotlib.pyplot as plt
import cudf, cupy, time, rmm
print( cudf.__version__ )

from dask.distributed import Client, wait
from dask_cuda import LocalCUDACluster
from dask.utils import parse_bytes
import cudf
import dask_cudf

import os
import time
import glob
import gc

import cupy as cp          # CuPy is an implementation of NumPy-compatible multi-dimensional array on GPU
import cudf                # cuDF is an implementation of Pandas-like Dataframe on GPU
import rmm                 # library for pre-allocating memory on GPU
import dask                # dask is an open-source library to nateively scale Python on multiple workers/nodes
import dask_cudf           # dask_cudf uses dask to scale cuDF dataframes on multiple workers/nodes

import numpy as np
import time
import xgboost as xgb
print( 'xgb', xgb.__version__ )

# More dask / dask_cluster related libraries to scale NVTabular
from dask_cuda import LocalCUDACluster
from dask.distributed import Client
from dask.distributed import wait
from dask.utils import parse_bytes
from dask.delayed import delayed

# NVTabular is the core library, we will use here for feature engineering/preprocessing on GPU
#import nvtabular as nvt
#from nvtabular.utils import device_mem_size
!rm -r dask-worker-space

import xgboost as xgb
print('XGB Version',xgb.__version__)

start = time.time()

from sklearn.metrics import auc
from sklearn.metrics import roc_auc_score
from sklearn.metrics import average_precision_score
from sklearn.metrics import log_loss

def compute_rce_fast(pred, gt):
    cross_entropy = log_loss(gt, pred)
    yt = np.mean(gt)     
    strawman_cross_entropy = -(yt*np.log(yt) + (1 - yt)*np.log(1 - yt))
    return (1.0 - cross_entropy/strawman_cross_entropy)*100.0

def convert_int32(df, feats=[]):
    for f in feats:
        df[f] = df[f].astype(np.int32)
    gc.collect()
    
def convert_int16(df, feats=[]):
    for f in feats:
        df[f] = df[f].astype(np.int16)
    gc.collect()
    
def convert_int8(df, feats=[]):
    for f in feats:
        df[f] = df[f].astype(np.int8)        
    gc.collect()

trainfiles = glob.glob('week3-valid/*.parquet' )
len(trainfiles), trainfiles

pd.read_parquet(trainfiles[0]).columns

%%time

train = pd.concat([pd.read_parquet(fn) for fn in trainfiles]).reset_index(drop=True)
train['decline'] = train['decline']//7.19
train['a_follower_count'] = train['a_follower_count'] // 819
train['a_following_count'] = train['a_following_count'] // 712
train['b_follower_count'] = train['b_follower_count'] // 819
train['b_following_count'] = train['b_following_count'] // 712
gc.collect()
train.shape

train.columns

%%time

dt = pd.read_parquet('TEMAPS_LOCAL/a_user_id-tweet_type.parquet')
features = list(dt.index.names)
train = train.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(train.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/b_user_id-tweet_type.parquet')
features = list(dt.index.names)
train = train.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(train.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/media-language-tweet_type-a_is_verified-b_is_verified-b_follows_a-tw_last_quest-decline-group.parquet')
features = list(dt.index.names)
train = train.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(train.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_original_user0-tweet_type.parquet')
features = list(dt.index.names)
train = train.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(train.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_original_user1-tweet_type.parquet')
features = list(dt.index.names)
train = train.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(train.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_word0-tweet_type.parquet')
features = list(dt.index.names)
train = train.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(train.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_rt_user0.parquet')
features = list(dt.index.names)
train = train.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(train.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/a_follower_count-a_following_count-b_follower_count-b_following_count-tweet_type-language-b_follows_a.parquet')
features = list(dt.index.names)
train = train.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(train.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/b_user_id.parquet')
features = list(dt.index.names)
train = train.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(train.shape)

%%time

valid = pd.read_parquet('/raid/kaggle/2021/recsys/input/validation.parquet')
valid['decline'] = valid['decline']//7.19
valid['a_follower_count'] = valid['a_follower_count'] // 819
valid['a_following_count'] = valid['a_following_count'] // 712
valid['b_follower_count'] = valid['b_follower_count'] // 819
valid['b_following_count'] = valid['b_following_count'] // 712
gc.collect()
valid.shape

%%time

dt = pd.read_parquet('TEMAPS_LOCAL/a_user_id-tweet_type.parquet')
features = list(dt.index.names)
valid = valid.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(valid.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/b_user_id-tweet_type.parquet')
features = list(dt.index.names)
valid = valid.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(valid.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/media-language-tweet_type-a_is_verified-b_is_verified-b_follows_a-tw_last_quest-decline-group.parquet')
features = list(dt.index.names)
valid = valid.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(valid.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_original_user0-tweet_type.parquet')
features = list(dt.index.names)
valid = valid.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(valid.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_original_user1-tweet_type.parquet')
features = list(dt.index.names)
valid = valid.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(valid.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_word0-tweet_type.parquet')
features = list(dt.index.names)
valid = valid.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(valid.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_rt_user0.parquet')
features = list(dt.index.names)
valid = valid.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(valid.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/a_follower_count-a_following_count-b_follower_count-b_following_count-tweet_type-language-b_follows_a.parquet')
features = list(dt.index.names)
valid = valid.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(valid.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/b_user_id.parquet')
features = list(dt.index.names)
valid = valid.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(valid.shape)

for f in [
'sums_te_reply-a_user_id-tweet_type',
       'counts_te_reply-a_user_id-tweet_type',
       'sums_te_retweet-a_user_id-tweet_type',
       'sums_te_retweet_comment-a_user_id-tweet_type',
       'sums_te_like-a_user_id-tweet_type',
       'sums_te_reply-b_user_id-tweet_type',
       'counts_te_reply-b_user_id-tweet_type',
       'sums_te_retweet-b_user_id-tweet_type',
       'sums_te_retweet_comment-b_user_id-tweet_type',
       'sums_te_like-b_user_id-tweet_type', 'multi_reply', 'multi_retweet',
       'multi_retweet_comment', 'multi_like', 'multi_counts', 'ouser0_reply',
       'ouser0_retweet', 'ouser0_retweet_comment', 'ouser0_like',
       'ouser0_counts', 'ouser1_reply', 'ouser1_retweet',
       'ouser1_retweet_comment', 'ouser1_like', 'ouser1_counts', 'word_reply',
       'word_retweet', 'word_retweet_comment', 'word_like', 'word_counts',
       'rtuser0_reply', 'rtuser0_retweet', 'rtuser0_retweet_comment',
       'rtuser0_like', 'rtuser0_counts', 'follow_reply', 'follow_retweet',
       'follow_retweet_comment', 'follow_like', 'follow_counts',
       'sums_b_user_reply', 'sums_b_user_retweet',
       'sums_b_user_retweet_comment', 'sums_b_user_like', 'counts_b_user_id'      
]:
    train[f] = train[f].astype(np.float32)
    valid[f] = valid[f].astype(np.float32)
gc.collect()

train.head()

targets = ['reply', 'retweet', 'retweet_comment', 'like']

features_base = [
    'media', 'tweet_type', 'language', 
    'a_follower_count', 'a_following_count', 'a_is_verified', 'a_account_creation',
    'b_follower_count', 'b_following_count', 'b_is_verified', 'b_account_creation', 'b_follows_a',
    'tw_len_token', 'tw_len_media', 'tw_len_photo', 'tw_len_video', 'tw_len_gif',
    'tw_len_quest', 'tw_count_capital_words', 'tw_count_excl_quest_marks', 'tw_count_special1',
    'tw_count_hash', 'tw_last_quest', 'tw_len_retweet', 'tw_len_rt', 'tw_count_at',
    'tw_count_words', 'tw_count_char',  'tw_rt_count_words', 'tw_rt_count_char',
    #'group',
    'dt_dow',
    #'dt_minute',
    'len_hashtags', 'len_links', 'len_domains', 'decline'
]

feature_reply = [
    
'sums_te_reply-a_user_id-tweet_type',
#'sums_te_retweet-a_user_id-tweet_type',
#'sums_te_retweet_comment-a_user_id-tweet_type',
#'sums_te_like-a_user_id-tweet_type',
'counts_te_reply-a_user_id-tweet_type',
    
'sums_te_reply-b_user_id-tweet_type',
#'sums_te_retweet-b_user_id-tweet_type',
#'sums_te_retweet_comment-b_user_id-tweet_type',
#'sums_te_like-b_user_id-tweet_type',
'counts_te_reply-b_user_id-tweet_type',
    
'multi_reply',
'multi_retweet',
'multi_retweet_comment',
'multi_like',
'multi_counts', 

'ouser0_reply',
'ouser0_retweet',
'ouser0_retweet_comment', 
'ouser0_like',
'ouser0_counts',

'ouser1_reply',
#'ouser1_retweet',
#'ouser1_retweet_comment', 
#'ouser1_like',
'ouser1_counts',

'word_reply',
'word_retweet',
'word_retweet_comment',
'word_like',
'word_counts',

'rtuser0_reply',
#'rtuser0_retweet',
#'rtuser0_retweet_comment',
#'rtuser0_like',
'rtuser0_counts',
    
'follow_reply',
'follow_retweet',
'follow_retweet_comment',
'follow_like',
'follow_counts',
    
'sums_b_user_reply',
'sums_b_user_retweet',
'sums_b_user_retweet_comment',
'sums_b_user_like',
'counts_b_user_id',
    
]

feature_reply

features = features_base+feature_reply
pickle.dump(features, open('xgbmodels-features-reply.pickle', 'wb'))
features

train[features].head()

train['fold'] = np.arange(train.shape[0]) % 5  #TRAIN['tweet_id'] % 5
print(train.groupby('fold')['media'].agg('count'))
gc.collect();gc.collect();gc.collect();

train['positive'] = 0
train.loc[(train['reply']>0)|(train['retweet']>0)|(train['retweet_comment']>0)|(train['like']>0), 'positive'] = 1
print(train['positive'].value_counts())

train['target'] = 0
train.loc[train['positive']==1,'target'] = 1
train.loc[train['reply']==1,'target'] = 2
print( train['target'].value_counts() )

#!rm -r giba_xgbmodels
#!mkdir giba_xgbmodels
xgb_parms = { 
    'max_depth': 8, 
    'learning_rate':0.04, 
    'subsample':0.50,
    'sampling_method': 'gradient_based',
    'colsample_bytree':0.80, 
    'grow_policy ': 'lossguide', #'depthwise', 'lossguide'
    'eval_metric':'auc',
    'objective':'binary:logistic',
    'nthread': 1,
    'tree_method':'gpu_hist',
    'predictor' : 'gpu_predictor'
}
print(xgb_parms)

target = 'reply'

feature_reply = [
    
'sums_te_reply-a_user_id-tweet_type',
#'sums_te_retweet-a_user_id-tweet_type',
#'sums_te_retweet_comment-a_user_id-tweet_type',
#'sums_te_like-a_user_id-tweet_type',
'counts_te_reply-a_user_id-tweet_type',
    
'sums_te_reply-b_user_id-tweet_type',
#'sums_te_retweet-b_user_id-tweet_type',
#'sums_te_retweet_comment-b_user_id-tweet_type',
#'sums_te_like-b_user_id-tweet_type',
'counts_te_reply-b_user_id-tweet_type',
    
'multi_reply',
'multi_retweet',
'multi_retweet_comment',
'multi_like',
'multi_counts', 

'ouser0_reply',
'ouser0_retweet',
'ouser0_retweet_comment', 
'ouser0_like',
'ouser0_counts',

'ouser1_reply',
#'ouser1_retweet',
#'ouser1_retweet_comment', 
#'ouser1_like',
'ouser1_counts',

'word_reply',
'word_retweet',
'word_retweet_comment',
'word_like',
'word_counts',

'rtuser0_reply',
#'rtuser0_retweet',
#'rtuser0_retweet_comment',
#'rtuser0_like',
'rtuser0_counts',
    
'follow_reply',
'follow_retweet',
'follow_retweet_comment',
'follow_like',
'follow_counts',
    
'sums_b_user_reply',
'sums_b_user_retweet',
'sums_b_user_retweet_comment',
'sums_b_user_like',
'counts_b_user_id',
    
]
features = features_base+feature_reply
pickle.dump(features, open('xgbmodels-features-reply.pickle', 'wb'))

print('TARGET:', target )
for group in range(5):
    start = time.time();
    dtrain = xgb.DMatrix(data=train.loc[(train.group==group), features], label=train.loc[(train.group==group), target] ) )
    gc.collect();gc.collect();gc.collect();
    print( time.time() - start, 's' )

    start = time.time(); print('Training...')
    model = xgb.train(
        xgb_parms, 
        dtrain=dtrain,
        #evals=[(dvalid2,'valid2')],
        num_boost_round=550,
        #early_stopping_rounds=100,
        verbose_eval=100,
    )
    pickle.dump(model, open('giba_xgbmodels/model_' + str(target) + '_' + str(group) + '.pickle', 'wb'))


xgb_parms = { 
    'max_depth': 8, 
    'learning_rate':0.04, 
    'subsample':0.50,
    'sampling_method': 'gradient_based',
    'colsample_bytree':0.80, 
    'grow_policy ': 'lossguide', #'depthwise', 'lossguide'
    'eval_metric':'auc',
    'objective':'binary:logistic',
    'nthread': 1,
    'tree_method':'gpu_hist',
    'predictor' : 'gpu_predictor'
}

target = 'retweet'

feature_retweet = [
    
#'sums_te_reply-a_user_id-tweet_type',
'sums_te_retweet-a_user_id-tweet_type',
#'sums_te_retweet_comment-a_user_id-tweet_type',
#'sums_te_like-a_user_id-tweet_type',
'counts_te_reply-a_user_id-tweet_type',
    
#'sums_te_reply-b_user_id-tweet_type',
'sums_te_retweet-b_user_id-tweet_type',
#'sums_te_retweet_comment-b_user_id-tweet_type',
#'sums_te_like-b_user_id-tweet_type',
'counts_te_reply-b_user_id-tweet_type',
    
'multi_reply',
'multi_retweet',
'multi_retweet_comment',
'multi_like',
'multi_counts', 

'ouser0_reply',
'ouser0_retweet',
'ouser0_retweet_comment', 
'ouser0_like',
'ouser0_counts',

#'ouser1_reply',
'ouser1_retweet',
#'ouser1_retweet_comment', 
#'ouser1_like',
'ouser1_counts',

'word_reply',
'word_retweet',
'word_retweet_comment',
'word_like',
'word_counts',

#'rtuser0_reply',
'rtuser0_retweet',
#'rtuser0_retweet_comment',
#'rtuser0_like',
'rtuser0_counts',
    
'follow_reply',
'follow_retweet',
'follow_retweet_comment',
'follow_like',
'follow_counts',
    
'sums_b_user_reply',
'sums_b_user_retweet',
'sums_b_user_retweet_comment',
'sums_b_user_like',
'counts_b_user_id',
    
]
features = features_base+feature_retweet
pickle.dump(features, open('xgbmodels-features-retweet.pickle', 'wb'))

print('TARGET:', target )
for group in range(5):
    start = time.time();
    dtrain = xgb.DMatrix(data=train.loc[(train.group==group), features], label=train.loc[(train.group==group), target] ) )
    gc.collect();gc.collect();gc.collect();
    print( time.time() - start, 's' )

    start = time.time(); print('Training...')
    model = xgb.train(
        xgb_parms, 
        dtrain=dtrain,
        #evals=[(dvalid2,'valid2')],
        num_boost_round=615,
        #early_stopping_rounds=100,
        verbose_eval=100,
    )
    pickle.dump(model, open('giba_xgbmodels/model_' + str(target) + '_' + str(group) + '.pickle', 'wb'))


xgb_parms = { 
    'max_depth': 8, 
    'learning_rate':0.02, 
    'subsample':0.50,
    'sampling_method': 'gradient_based',
    'colsample_bytree':0.80, 
    'grow_policy ': 'lossguide', #'depthwise', 'lossguide'
    'eval_metric':'auc',
    'objective':'binary:logistic',
    'nthread': 1,
    'tree_method':'gpu_hist',
    'predictor' : 'gpu_predictor'
}

target = 'retweet_comment'

feature_retweet_comment = [
    
#'sums_te_reply-a_user_id-tweet_type',
#'sums_te_retweet-a_user_id-tweet_type',
'sums_te_retweet_comment-a_user_id-tweet_type',
#'sums_te_like-a_user_id-tweet_type',
'counts_te_reply-a_user_id-tweet_type',
    
#'sums_te_reply-b_user_id-tweet_type',
#'sums_te_retweet-b_user_id-tweet_type',
'sums_te_retweet_comment-b_user_id-tweet_type',
#'sums_te_like-b_user_id-tweet_type',
'counts_te_reply-b_user_id-tweet_type',
    
'multi_reply',
'multi_retweet',
'multi_retweet_comment',
'multi_like',
'multi_counts', 

'ouser0_reply',
'ouser0_retweet',
'ouser0_retweet_comment', 
'ouser0_like',
'ouser0_counts',

#'ouser1_reply',
#'ouser1_retweet',
'ouser1_retweet_comment', 
#'ouser1_like',
'ouser1_counts',

'word_reply',
'word_retweet',
'word_retweet_comment',
'word_like',
'word_counts',

#'rtuser0_reply',
#'rtuser0_retweet',
'rtuser0_retweet_comment',
#'rtuser0_like',
'rtuser0_counts',
    
'follow_reply',
'follow_retweet',
'follow_retweet_comment',
'follow_like',
'follow_counts',
    
'sums_b_user_reply',
'sums_b_user_retweet',
'sums_b_user_retweet_comment',
'sums_b_user_like',
'counts_b_user_id',
    
]
features = features_base+feature_retweet_comment
pickle.dump(features, open('xgbmodels-features-retweet_comment.pickle', 'wb'))

print('TARGET:', target )
for group in range(5):
    start = time.time();
    dtrain = xgb.DMatrix(data=train.loc[(train.group==group), features], label=train.loc[(train.group==group), target] ) )
    gc.collect();gc.collect();gc.collect();
    print( time.time() - start, 's' )

    start = time.time(); print('Training...')
    model = xgb.train(
        xgb_parms, 
        dtrain=dtrain,
        #evals=[(dvalid2,'valid2')],
        num_boost_round=800,
        #early_stopping_rounds=100,
        verbose_eval=100,
    )
    pickle.dump(model, open('giba_xgbmodels/model_' + str(target) + '_' + str(group) + '.pickle', 'wb'))


xgb_parms = { 
    'max_depth': 8, 
    'learning_rate':0.04, 
    'subsample':0.50,
    'sampling_method': 'gradient_based',
    'colsample_bytree':0.80, 
    'grow_policy ': 'lossguide', #'depthwise', 'lossguide'
    'eval_metric':'auc',
    'objective':'binary:logistic',
    'nthread': 1,
    'tree_method':'gpu_hist',
    'predictor' : 'gpu_predictor'
}

target = 'like'

feature_like = [
    
#'sums_te_reply-a_user_id-tweet_type',
#'sums_te_retweet-a_user_id-tweet_type',
#'sums_te_retweet_comment-a_user_id-tweet_type',
'sums_te_like-a_user_id-tweet_type',
'counts_te_reply-a_user_id-tweet_type',
    
#'sums_te_reply-b_user_id-tweet_type',
#'sums_te_retweet-b_user_id-tweet_type',
#'sums_te_retweet_comment-b_user_id-tweet_type',
'sums_te_like-b_user_id-tweet_type',
'counts_te_reply-b_user_id-tweet_type',
    
'multi_reply',
'multi_retweet',
'multi_retweet_comment',
'multi_like',
'multi_counts', 

'ouser0_reply',
'ouser0_retweet',
'ouser0_retweet_comment', 
'ouser0_like',
'ouser0_counts',

#'ouser1_reply',
#'ouser1_retweet',
#'ouser1_retweet_comment', 
'ouser1_like',
'ouser1_counts',

'word_reply',
'word_retweet',
'word_retweet_comment',
'word_like',
'word_counts',

#'rtuser0_reply',
#'rtuser0_retweet',
#'rtuser0_retweet_comment',
'rtuser0_like',
'rtuser0_counts',
    
'follow_reply',
'follow_retweet',
'follow_retweet_comment',
'follow_like',
'follow_counts',
    
'sums_b_user_reply',
'sums_b_user_retweet',
'sums_b_user_retweet_comment',
'sums_b_user_like',
'counts_b_user_id',
    
]
features = features_base+feature_like
pickle.dump(features, open('xgbmodels-features-like.pickle', 'wb'))

print('TARGET:', target )
for group in range(5):
    start = time.time();
    dtrain = xgb.DMatrix(data=train.loc[(train.group==group), features], label=train.loc[(train.group==group), target] ) )
    gc.collect();gc.collect();gc.collect();
    print( time.time() - start, 's' )

    start = time.time(); print('Training...')
    model = xgb.train(
        xgb_parms, 
        dtrain=dtrain,
        #evals=[(dvalid2,'valid2')],
        num_boost_round=694,
        #early_stopping_rounds=100,
        verbose_eval=100,
    )
    pickle.dump(model, open('giba_xgbmodels/model_' + str(target) + '_' + str(group) + '.pickle', 'wb'))


!zip -r xgb.zip giba_xgbmodels/*

pickle.load(open('xgbmodels-features-like.pickle', 'rb'))
