==> RecSys2021_Challenge/01_training/stage1/giba/4-stacking.ipynb <==
# Copyright 2021 NVIDIA CORPORATION
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

import pandas as pd
import numpy as np
from sklearn.metrics import average_precision_score, roc_auc_score
import gc
import glob
from tqdm.notebook import tqdm
import time
import pickle
import xgboost as xgb
print( 'xgb', xgb.__version__ )


def compute_rce_fast(pred, gt):
    yt = np.mean(gt)     
    strawman_cross_entropy = -(yt*np.log(yt) + (1 - yt)*np.log(1 - yt))
    
    cross_entropy = -(gt*np.log(pred+1e-6) + (1 - gt)*np.log(1 - pred + 1e-6))
    cross_entropy = np.mean(cross_entropy)
    #print( cross_entropy, strawman_cross_entropy )
    return (1.0 - cross_entropy/strawman_cross_entropy)*100.0

gc.collect()

dt = pd.read_csv('stacking_models/part-00000_targets', sep='\x01', header=None)
print( dt.shape )
all_features = [
    'text_tokens',    ###############
    'hashtags',       #Tweet Features
    'tweet_id_org',       #
    'media',          #
    'links',          #
    'domains',        #
    'tweet_type',     #
    'language',       #
    'timestamp',      ###############
    'a_user_id',              ###########################
    'a_follower_count',       #Engaged With User Features
    'a_following_count',      #
    'a_is_verified',          #
    'a_account_creation',     ###########################
    'b_user_id_org',              #######################
    'b_follower_count',       #Engaging User Features
    'b_following_count',      #
    'b_is_verified',          #
    'b_account_creation',     #######################
    'b_follows_a',    #################### Engagement Features
    
    'reply',
    'retweet',
    'retweet_comment',
    'like',
]
dt.columns = all_features

# dt['group'] = 0
# dt['group'] = dt['group'] + 1*(dt['a_follower_count']>=240)
# dt['group'] = dt['group'] + 1*(dt['a_follower_count']>=588)
# dt['group'] = dt['group'] + 1*(dt['a_follower_count']>=1331)
# dt['group'] = dt['group'] + 1*(dt['a_follower_count']>=3996)
# dt['group'] = dt['group'].astype(np.int8)

for tgt in ['reply', 'retweet', 'retweet_comment', 'like']:
    dt[tgt] = dt[tgt].fillna(0.)
    dt[tgt] = 1*(dt[tgt]>0)

for tgt in [['reply','retweet','retweet_comment','like']]:
    dt[tgt] = dt[tgt].astype(np.float32)
    
dt = dt[['tweet_id_org', 'b_user_id_org', 'reply', 'retweet', 'retweet_comment', 'like' ]]
# gc.collect()
#dt = dt.sort_values(['tweet_id', 'b_user_id']).reset_index(drop=True)
#print( dt[['reply', 'retweet', 'retweet_comment', 'like']].mean() )
#print( dt.shape )
#gc.collect()
dt.head()

features = ['tweet_id_org', 'b_user_id_org',
    #'reply', 'retweet', 'retweet_comment', 'like', 
    'text',
    'media', 'tweet_type',
    'language', 'timestamp', 
    'a_user_id', 'a_user_id32', 'a_follower_count', 'a_following_count', 'a_is_verified', 'a_account_creation', 
    'b_user_id', 'b_user_id32', 'b_follower_count', 'b_following_count', 'b_is_verified', 'b_account_creation', 'b_follows_a',
    
    'tw_len_token', 'tw_len_media', 'tw_len_photo', 'tw_len_video', 'tw_len_gif',
    'tw_len_quest', 'tw_count_capital_words', 'tw_count_excl_quest_marks',
    'tw_count_special1', 'tw_count_hash', 'tw_last_quest', 'tw_len_retweet',
    'tw_len_rt', 'tw_count_at', 'tw_count_words', 'tw_count_char',
    'tw_rt_count_words', 'tw_rt_count_char', 'len_hashtags', 'len_links', 'len_domains', 'decline',
    
    'tw_original_user0', 'tw_original_user1', 'tw_original_user2',
    'tw_rt_user0',
    'tw_word0', 'tw_word1',
    'tw_tweet',
    
    'group', 'dt_day', 'dt_dow', 'dt_minute',
]
gc.collect()

TRAIN = []
for files in glob.glob('stacking_models/part*.parquet'):
    TRAIN.append(pd.read_parquet(files, columns=features))
TRAIN = pd.concat(TRAIN)
print(TRAIN.shape)
print(TRAIN.columns)

TRAIN = TRAIN.merge(dt, on=['tweet_id_org', 'b_user_id_org'], how='left')
TRAIN.head()

targets = ['reply', 'retweet', 'retweet_comment', 'like']
TRAIN.groupby('group')[targets].agg('mean')

TRAIN.head()

files = [
    'results_benny_nn.csv',
    'results_benny_xgb.csv',
    'results-chris-nn.csv',
    'results-chris-xgb.csv',
    'results-giba-xgb.csv',
    'results_bo.csv',    
]

for n, fn in tqdm(enumerate(files)):
    dt = pd.read_csv('stacking_models/'+fn, header=None)
    dt.columns = ['tweet_id_org', 'b_user_id_org', 'reply_'+str(n), 'retweet_'+str(n), 'retweet_comment_'+str(n), 'like_'+str(n)]
    for tgt in ['reply_'+str(n),'retweet_'+str(n),'retweet_comment_'+str(n),'like_'+str(n)]:
        dt[tgt] = dt[tgt].astype(np.float32)
    TRAIN = TRAIN.merge(dt, on=['tweet_id_org', 'b_user_id_org'], how='left')
    print(dt.shape, TRAIN.shape)
    del dt
    _ = gc.collect()

TRAIN.head()

TRAIN.to_parquet('stacking_models/stack.parquet')

TRAIN = pd.read_parquet('stacking_models/stack.parquet')
TRAIN['fold'] = np.arange(TRAIN.shape[0])%5
TRAIN.shape

TRAIN.columns

%%time
dt = pd.read_parquet('TEMAPS_LOCAL/te-a_user_id32-b_user_id32_count.parquet')
TRAIN = TRAIN.merge(dt, left_on=['a_user_id32','b_user_id32'], right_index=True, how='left')
gc.collect()

dt = pd.read_parquet('TEMAPS_LOCAL/a_user_id-tweet_type.parquet')
features = list(dt.index.names)
TRAIN = TRAIN.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(TRAIN.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/b_user_id-tweet_type.parquet')
features = list(dt.index.names)
TRAIN = TRAIN.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(TRAIN.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/media-language-tweet_type-a_is_verified-b_is_verified-b_follows_a-tw_last_quest-decline-group.parquet')
features = list(dt.index.names)
TRAIN = TRAIN.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(TRAIN.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_original_user0-tweet_type.parquet')
features = list(dt.index.names)
TRAIN = TRAIN.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(TRAIN.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_original_user1-tweet_type.parquet')
features = list(dt.index.names)
TRAIN = TRAIN.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(TRAIN.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_word0-tweet_type.parquet')
features = list(dt.index.names)
TRAIN = TRAIN.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(TRAIN.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/tw_rt_user0.parquet')
features = list(dt.index.names)
TRAIN = TRAIN.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(TRAIN.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/a_follower_count-a_following_count-b_follower_count-b_following_count-tweet_type-language-b_follows_a.parquet')
features = list(dt.index.names)
TRAIN = TRAIN.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(TRAIN.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/b_user_id.parquet')
features = list(dt.index.names)
TRAIN = TRAIN.merge(dt, left_on=features, right_index=True, how='left')
del dt; gc.collect()
print(TRAIN.shape)

dt = pd.read_parquet('TEMAPS_LOCAL/b_user_id.parquet').reset_index()
dt.columns = ['a_user_id','t0','t1','t2','t3','t4']
TRAIN = TRAIN.merge(dt, on=['a_user_id'], how='left')

TRAIN.head()

targets = ['reply', 'retweet', 'retweet_comment', 'like']

features_base = [
    'media', 'tweet_type', 'language', 
    'a_follower_count', 'a_following_count', 'a_is_verified', 'a_account_creation',
    'b_follower_count', 'b_following_count', 'b_is_verified', 'b_account_creation', 'b_follows_a',
    'tw_len_token', 'tw_len_media', 'tw_len_photo', 'tw_len_video', 'tw_len_gif',
    'tw_len_quest', 'tw_count_capital_words', 'tw_count_excl_quest_marks', 'tw_count_special1',
    'tw_count_hash', 'tw_last_quest', 'tw_len_retweet', 'tw_len_rt', 'tw_count_at',
    'tw_count_words', 'tw_count_char',  'tw_rt_count_words', 'tw_rt_count_char',
    'dt_dow',
    'dt_minute',
    'len_hashtags', 'len_links', 'len_domains', 'decline'
]

TRAIN['count1'] = TRAIN.groupby(['a_user_id'])['dt_dow'].transform('count')
TRAIN['count2'] = TRAIN.groupby(['b_user_id'])['dt_dow'].transform('count')
TRAIN['count3'] = TRAIN.groupby(['tw_tweet'])['dt_dow'].transform('count')
TRAIN['count4'] = TRAIN.groupby(['a_user_id','b_user_id'])['dt_dow'].transform('count')

TRAIN.head()

train0 = TRAIN.loc[TRAIN.fold!=0].copy()
train1 = TRAIN.loc[TRAIN.fold==0].copy()

train0.shape, train1.shape

feature_s1 = [
    
    'reply_0',
    'retweet_0',
    'retweet_comment_0',
    'like_0',

    'reply_1',
    'retweet_1',
    'retweet_comment_1',
    'like_1',

    'reply_2',
    'retweet_2',
    'retweet_comment_2',
    'like_2',

    'reply_3',
    'retweet_3',
    'retweet_comment_3',
    'like_3',

    'reply_4',
    'retweet_4',
    'retweet_comment_4',
    'like_4',

    'reply_5',
    'retweet_5',
    'retweet_comment_5',
    'like_5',
    
    'count1', 'count2', 'count3', 'count4',
    
   'sums_te_reply-a_user_id-tweet_type',
   'counts_te_reply-a_user_id-tweet_type',
   'sums_te_retweet-a_user_id-tweet_type',
   'sums_te_retweet_comment-a_user_id-tweet_type',
   'sums_te_like-a_user_id-tweet_type',
   'sums_te_reply-b_user_id-tweet_type',
   'counts_te_reply-b_user_id-tweet_type',
   'sums_te_retweet-b_user_id-tweet_type',
   'sums_te_retweet_comment-b_user_id-tweet_type',
   'sums_te_like-b_user_id-tweet_type', 'multi_reply', 'multi_retweet',
   'multi_retweet_comment', 'multi_like', 'multi_counts', 'ouser0_reply',
   'ouser0_retweet', 'ouser0_retweet_comment', 'ouser0_like',
   'ouser0_counts', 'ouser1_reply', 'ouser1_retweet',
   'ouser1_retweet_comment', 'ouser1_like', 'ouser1_counts', 'word_reply',
   'word_retweet', 'word_retweet_comment', 'word_like', 'word_counts',
   'rtuser0_reply', 'rtuser0_retweet', 'rtuser0_retweet_comment',
   'rtuser0_like', 'rtuser0_counts', 'follow_reply', 'follow_retweet',
   'follow_retweet_comment', 'follow_like', 'follow_counts',
   'sums_b_user_reply', 'sums_b_user_retweet',
   'sums_b_user_retweet_comment', 'sums_b_user_like', 'counts_b_user_id',
   't0', 't1', 't2', 't3', 't4',    
    
    'ab_count',
    
]
features = features_base + feature_s1
features

def calc_rce(df, target='reply'):
    SC = []
    for g in range(5):
        SC.append( compute_rce_fast(df.loc[df.group==g,'y_'+target].values,  df.loc[df.group==g,target].values ) )    
    print(SC)
    return np.mean(SC)

def calc_ap(df, target='reply'):
    SC = []
    for g in range(5):
        SC.append( average_precision_score(df.loc[df.group==g,target].values,  df.loc[df.group==g,'y_'+target].values ) )    
    print(SC)
    return np.mean(SC)

LR = {
    'reply': 0.0375,
    'retweet': 0.063,
    'retweet_comment': 0.021,
    'like': 0.113625,
}

fold = 0

SC1 = []
SC2 = []
for target in ['reply', 'retweet', 'retweet_comment', 'like']:
    print('TARGET:', target )
    xgb_parms = { 
        'max_depth': 9, 
        'learning_rate': LR[target]/1, 
        'subsample':0.50,
        #'sampling_method': 'gradient_based',
        'max_bin': 256,
        'colsample_bytree':0.80,
        'grow_policy ': 'depthwise', #'depthwise', 'lossguide'
        'eval_metric':['logloss','auc'],
        'objective':'binary:logistic',
        'nthread': 1,
        'tree_method':'gpu_hist',
        'predictor' : 'gpu_predictor',
    }

    ROUND = []
    #for group in range(5):
    start = time.time();

    #W = train0['group'].values.copy()
    #grp, cnt = np.unique(W,return_counts=True)
    #cnt = cnt[0] / cnt
    #cnt = np.log1p(cnt)
    #W = cnt[W]
    #ind = train0['positive'].values == 0
    #W[ind] = W[ind]/5.
    #print( np.unique(W, return_counts=True) )

    #FW = np.ones(len(features)) 
    #FW[:30] = 2
    #FW[-5:] = 2
    #FW = FW / 2.
    #print( np.unique(FW, return_counts=True) )

    dtrain = xgb.DMatrix(data=TRAIN[features], label=TRAIN[target] )#, weight=W  )
    #dtrain.set_info(feature_weights=FW)
    dvalid = xgb.DMatrix(data=train1[features], label=train1[target] )#, weight=W  )
    gc.collect();gc.collect();gc.collect();
    print( time.time() - start, 's' )

    start = time.time(); print('Training...')
    model = xgb.train(
        xgb_parms, 
        dtrain=dtrain,
        evals=[(dvalid,'valid')],
        num_boost_round=1000,
        #early_stopping_rounds=100,
        verbose_eval=10,
    )
    train1['y_'+target] = model.predict(dvalid)

    sc1 = calc_rce(train1, target)
    sc2 = calc_ap(train1, target)
    print( 'RCE:', sc1, sc2 )
    SC1.append(sc1)
    SC2.append(sc2)

    #sc2 = compute_rce_fast(ypred1, train1[target].values )
    #ap2 = average_precision_score(train1[target].values, ypred1)
    #SC1.append(sc2)
    #SC2.append(ap2)
    # ROUND.append(model.best_iteration)
    # print( model.best_iteration, 'RCE:', sc1, sc2, ap1, ap2, int(time.time() - start), 's' )
    # print()
    pickle.dump(model, open('giba_xgbmodels/model_' + str(target) + '_' + str(fold) + '.pickle', 'wb'))

    # del dtrain, dvalid2, model
    gc.collect()
    #break

    #print('RCE:', np.mean(SC1), np.mean(SC2) )
    print()
    #print()
print(np.sum(SC1), np.sum(SC2))

imp = pd.DataFrame().from_dict(model.get_score(importance_type='gain'), orient='index').sort_values(0, ascending=False).reset_index()
imp.head(40)

imp.tail(20)
